third_party/blink/web_tests/external/wpt/screen-orientation/unlock-expected.txt

This is a testharness.js-based test.
PASS unlock() doesn't throw when there is no lock
PASS unlock() returns a void value
PASS unlock() doesn't throw when there is no lock with fullscreen
PASS unlock() aborts a pending lock request
FAIL unlock() aborts a pending lock request across documents assert_unreached: Should have rejected: undefined Reached unreachable code
Harness: the test ran to completion.
